requirements.txt
selenium
webdriver-manager
python-dotenv
flask
flask-cors
